+++ ilyrics/scripts/powershell_f/check_player.ps1
@@ -1,1 +1,2 @@
 # WORKING ON WINDOWS VERSION!
+# WATCH US ON GITHUB! github.com/Mondonno/iLyrics