+++ ilyrics_py/main.py
@@ -1,21 +1,23 @@
 import os
-
-##  _ _                _          
-## (_) |              (_)         
-##  _| |    _   _ _ __ _  ___ ___ 
-## | | |   | | | | '__| |/ __/ __|
-## | | |___| |_| | |  | | (__\__ \
-## |_|______\__, |_|  |_|\___|___/
-##           __/ |                
-##          |___/                 
-## ------------------------------------------>
-## iLyrics - Get every played song lirycs!
-## Author - github.com/Mondonno
-## Github Repository - github.com/Mondonno/iLyrics
-## Created with LOVE
-## ------------------------------------------>
-## This is the main file of the iLyrics python rewrite
-
+import requests
+  
+"""
+    _ _                _          
+   (_) |              (_)         
+    _| |    _   _ _ __ _  ___ ___ 
+   | | |   | | | | '__| |/ __/ __|
+   | | |___| |_| | |  | | (__\__ \\
+   |_|______\__, |_|  |_|\___|___/
+             __/ |                
+           |___/                 
+   ------------------------------------------>
+   iLyrics - Get every played song lirycs!
+   Author - github.com/Mondonno
+   Github Repository - github.com/Mondonno/iLyrics
+   Created with LOVE
+   ------------------------------------------>
+   This is the main file of the iLyrics python rewrite
+"""
 actual_song = ""
 checking_time = 5000