+++ CHANGELOG.md
@@ -4,4 +4,4 @@ v1.0
    - Spotify
    - Quick time player
  - New GENIUS api
- - Used and optimized the ps and osa scripts 
+ - Used and optimized the PS and APS scripts